feat: split rolldown-generic core into @sanity/vanilla-extract-rolldown-plugin
The new package contains everything that only needs rolldown: .css.ts
compilation with plugin hook filters, single-file lightningcss-optimized
CSS extraction, and the inject wiring (relative import, nodeCompat
self-referential import + no-op shim emission). Host tools provide
resolved defaults (target, packageName, cwd) through api.setBuildContext().

@sanity/vanilla-extract-tsdown-plugin is now a thin tsdown adapter with an
unchanged public API: it forwards tsdown's resolved config through
tsdownConfigResolved and keeps writing the conditional CSS export through
exports.customExports in tsdownConfig.
