+++ packages/@sanity/vanilla-extract-rolldown-plugin/test/fixtures/dynamic/index.ts
