+++ packages/@sanity/vanilla-extract-rolldown-plugin/test/fixtures/dynamic/lazy.css.ts
