+++ packages/@sanity/vanilla-extract-rolldown-plugin/test/fixtures/dynamic/static.css.ts
