+++ packages/@sanity/vanilla-extract-rolldown-plugin/tsconfig.dist.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["./tsconfig", "@sanity/tsconfig/isolated-declarations"],
+  "include": ["./src"]
+}
