+++ packages/@sanity/vanilla-extract-rolldown-plugin/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "@sanity/tsconfig/strictest",
+  "exclude": ["**/dist/**"],
+  "compilerOptions": {
+    "rootDir": "."
+  }
+}
