+++ packages/@sanity/vanilla-extract-rolldown-plugin/tsdown.config.ts
@@ -1,0 +1,6 @@
+import {defineConfig} from '@sanity/tsdown-config'
+
+export default defineConfig({
+  tsconfig: 'tsconfig.dist.json',
+  platform: 'node',
+})
