+++ packages/@sanity/vanilla-extract-rolldown-plugin/vitest.config.ts
@@ -1,0 +1,7 @@
+import {defineConfig} from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    typecheck: {enabled: true},
+  },
+})
