+++ packages/@sanity/vanilla-extract-tsdown-plugin/test/fixtures/basic/index.ts
@@ -1,10 +1,9 @@
-import {getButtonClassName} from './button.ts'
 import {box} from './styles.css.ts'
 
 /**
- * Returns the vanilla-extract generated class names. The styles themselves are extracted into a
+ * Returns the vanilla-extract generated class name. The styles themselves are extracted into a
  * separate CSS asset by the plugin.
  */
-export function getClassNames(): string {
-  return `${getButtonClassName()} ${box}`
+export function getClassName(): string {
+  return box
 }
