+++ tsconfig.json
@@ -4,6 +4,7 @@
     {"path": "./packages/@sanity/parse-package-json"},
     {"path": "./packages/@sanity/pkg-utils"},
     {"path": "./packages/@sanity/tsdown-config"},
+    {"path": "./packages/@sanity/vanilla-extract-rolldown-plugin"},
     {"path": "./packages/@sanity/vanilla-extract-tsdown-plugin"}
   ]
 }
